Move moved messageData objects to its own files
