+++ Source/Model/Message/ZMClientMessage+GenericMessage.swift
@@ -90,42 +90,3 @@ extension ZMClientMessage {
         }
     }
 }
-
-// MARK: Message data
-extension ZMClientMessage: ZMKnockMessageData {
-    
-    public override var textMessageData: ZMTextMessageData? {
-        let isTextMessage = self.genericMessage?.textData != nil
-        return isTextMessage ? self : nil
-    }
-    
-    public override var imageMessageData: ZMImageMessageData? {
-        return nil
-    }
-    
-    public override var knockMessageData: ZMKnockMessageData? {
-        let isKnockMessage = self.genericMessage?.knockData != nil
-        return isKnockMessage ? self : nil
-    }
-    
-    public override var fileMessageData: ZMFileMessageData? {
-        return nil
-    }
-    
-    public override var locationMessageData: LocationMessageData? {
-        switch underlyingMessage?.content {
-        case .location(_)?:
-            return self
-        case .ephemeral(let data)?:
-            switch data.content {
-            case .location(_)?:
-                return self
-            default:
-                return nil
-            }
-            
-        default:
-            return nil
-        }
-    }
-}
